[CI-4083] Refactor util functions to split up in modules by responsibility, fix imports, add util functions to read custom ui variant from session storage and apply logic based on it
