+++ src/components/Autocomplete/SectionItemsList/SectionItemsList.tsx
@@ -1,7 +1,8 @@
 import React, { ReactElement, useContext } from 'react';
 import { Section } from '../../../types';
 import SectionItem from '../SectionItem/SectionItem';
-import { camelToStartCase, translate } from '../../../utils';
+import { translate } from '../../../utils/helpers';
+import { camelToStartCase } from '../../../utils/format';
 import { CioAutocompleteContext } from '../CioAutocompleteProvider';
 import NoResults from '../AutocompleteResults/NoResults';
