+++ src/stories/Autocomplete/Component/AdvancedParameters.stories.tsx
@@ -1,6 +1,6 @@
 import { CioAutocomplete } from '../../../index';
 import { argTypes } from '../argTypes';
-import { stringifyWithDefaults } from '../../../utils';
+import { stringifyWithDefaults } from '../../../utils/format';
 import { ComponentTemplate, addComponentStoryDescription } from '.';
 import {
   advancedParametersDescription,
