+++ src/stories/tests/HooksTests.stories.tsx
@@ -2,7 +2,8 @@ import { within, userEvent } from '@storybook/testing-library';
 import { expect } from '@storybook/jest';
 import { CioAutocomplete } from '../../index';
 import { argTypes } from '../Autocomplete/argTypes';
-import { getCioClient, isTrackingRequestSent, sleep } from '../../utils';
+import { getCioClient, sleep } from '../../utils/helpers';
+import { isTrackingRequestSent } from '../../utils/tracking';
 import { HooksTemplate } from '../Autocomplete/Hook/index';
 import { apiKey, onSubmitDefault as onSubmit } from '../../constants';
 import { CioAutocompleteProps } from '../../types';
