+++ src/utils/helpers.ts
@@ -1,91 +1,18 @@
 import ConstructorIOClient from '@constructor-io/constructorio-client-javascript';
-import {
-  AutocompleteRequestType,
-  Nullable,
-  ConstructorClientOptions,
-} from '@constructor-io/constructorio-client-javascript/lib/types';
-// eslint-disable-next-line import/no-cycle
-import {
-  storageSetItem,
-  storageRemoveItem,
-  storeRecentSearch,
-  storeRecentAction,
-  CONSTANTS,
-} from './beaconUtils';
-import { isRecommendationsSection } from './typeGuards';
-import { Item, Section, UserDefinedSection, SectionsData, Translations, PodData } from './types';
-import version from './version';
+import { ConstructorClientOptions } from '@constructor-io/constructorio-client-javascript/lib/types';
+import { Item, Section, UserDefinedSection, SectionsData, Translations, PodData } from '../types';
+import version from '../version';
 
 export type GetItemPosition = (args: { item: Item; items: Item[] }) => {
   index: number;
   sectionId: string;
 };
 
-export function getFeatures(request: Partial<AutocompleteRequestType>) {
-  let featureDisplaySearchSuggestionImages = false;
-  let featureDisplaySearchSuggestionResultCounts = false;
-  let featureDisplayZeroStateRecommendations = true;
-
-  if (request?.features?.custom_autosuggest_ui === true) {
-    switch (request?.feature_variants?.custom_autosuggest_ui) {
-      case 'custom_autosuggest_ui_result_count':
-        featureDisplaySearchSuggestionResultCounts = true;
-        break;
-      case 'custom_autosuggest_ui_image':
-        featureDisplaySearchSuggestionImages = true;
-        break;
-      case 'custom_autosuggest_ui_image_result_count':
-        featureDisplaySearchSuggestionImages = true;
-        featureDisplaySearchSuggestionResultCounts = true;
-        break;
-      case 'custom_autosuggest_ui_disable_recommendations_in_zero_state':
-        featureDisplayZeroStateRecommendations = false;
-        break;
-      default:
-        break;
-    }
-  }
-
-  return {
-    featureDisplaySearchSuggestionImages,
-    featureDisplaySearchSuggestionResultCounts,
-    featureDisplayZeroStateRecommendations,
-  };
-}
-
 export const getItemPosition: GetItemPosition = ({ item, items }) => {
   const index = items.findIndex((itemInFlatList) => itemInFlatList?.id === item?.id);
   const sectionId = items[index]?.section;
   return { sectionId, index };
 };
-
-type CamelToStartCase = (camelCaseString: string) => string;
-
-export const camelToStartCase: CamelToStartCase = (camelCaseString) =>
-  camelCaseString
-    // insert a space between lower & upper
-    .replace(/([a-z])([A-Z])/g, '$1 $2')
-    // space before last upper in a sequence followed by lower
-    .replace(/\b([A-Z]+)([A-Z])([a-z])/, '$1 $2$3')
-    // uppercase the first letter
-    .replace(/([a-zA-Z])/, (str) => str.toUpperCase());
-
-export const toKebabCase = (str: string): string =>
-  str
-    .replace(/([A-Z])([A-Z][a-z])/g, '$1-$2')
-    .replace(/([a-z])([A-Z])/g, '$1-$2')
-    .replace(/[\s_]+/g, '-')
-    .toLowerCase();
-
-export function isTrackingRequestSent(trackingRequestUrl: string) {
-  // eslint-disable-next-line
-  const trackingRequestsQueue = window.localStorage?._constructorio_requests;
-
-  return (
-    trackingRequestsQueue &&
-    JSON.parse(trackingRequestsQueue)?.some((request) => request?.url?.includes(trackingRequestUrl))
-  );
-}
 
 export function clearConstructorRequests() {
   // eslint-disable-next-line
@@ -130,29 +57,6 @@ export const functionStrings = {
                           <p>{item.data?.price}</p>
                       </div>
                     )`,
-};
-
-export const stringifyWithDefaults = (obj) => {
-  // Stringify non-function values normally. Add a template block for functions to be replaced later
-  let res = JSON.stringify(
-    obj,
-    (key, value) => (value instanceof Function ? `${key}_CODE` : value),
-    '  '
-  );
-
-  // Replace template blocks with function strings
-  Array.from(res.matchAll(/"(\w*)_CODE"/g)).forEach((match) => {
-    const [codePlaceholder, key] = match;
-    const functionString = functionStrings[key];
-
-    if (functionString) {
-      res = res.replaceAll(codePlaceholder, functionString);
-    } else {
-      console.error(`Function string for ${key} not found.`); // eslint-disable-line
-    }
-  });
-
-  return res;
 };
 
 export const disableStoryActions = (story) => {
@@ -237,34 +141,6 @@ export const getActiveSectionsWithData = (
 
 export const escapeRegExp = (string: string) => string.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
 
-export const trackRecommendationView = (
-  target: HTMLElement,
-  activeSectionsWithData: Section[],
-  cioClient: Nullable<ConstructorIOClient>
-) => {
-  if (target.dataset.cnstrcRecommendationsPodId) {
-    // Pull recommendations from activeSectionsWithData by podId surfaced on target
-    const recommendationSection = activeSectionsWithData.find(
-      (section) =>
-        isRecommendationsSection(section) &&
-        section.podId === target.dataset.cnstrcRecommendationsPodId
-    );
-    const recommendationItems = recommendationSection?.data.map((item) => ({
-      itemId: item.data?.id,
-      itemName: item.value,
-      variationId: item.data?.variation_id,
-    }));
-
-    cioClient?.tracker.trackRecommendationView({
-      podId: target.dataset.cnstrcRecommendationsPodId,
-      numResultsViewed: recommendationItems?.length || 0,
-      url: window.location.href,
-      section: target.dataset.cnstrcSection,
-      items: recommendationItems,
-    });
-  }
-};
-
 export const getItemsForActiveSections = (activeSectionsWithData: Section[]) => {
   const items: Item[] = [];
   activeSectionsWithData?.forEach((config: Section) => {
@@ -285,20 +161,6 @@ export const translate = (word: string, translations?: Translations) => {
   if (translations) return translations[word];
 
   return localTranslations[word] || word;
-};
-
-export const trackSearchSubmit = (cioClient, term, autocompleteData = {}) => {
-  cioClient?.tracker.trackSearchSubmit(term, autocompleteData);
-  storageSetItem(CONSTANTS.SEARCH_TERM_STORAGE_KEY, term);
-  storeRecentSearch(term, {});
-  storeRecentAction(CONSTANTS.SEARCH_SUBMIT);
-};
-
-export const trackAutocompleteSelect = (cioClient, itemName, autocompleteData: any = {}) => {
-  cioClient?.tracker.trackAutocompleteSelect(itemName, autocompleteData);
-  if (autocompleteData?.section! === 'Products') {
-    storageRemoveItem(CONSTANTS.SEARCH_TERM_STORAGE_KEY);
-  }
 };
 
 export const logger = (error: any) => {
